requirements.txt
Brlapi==0.5.6
Django==1.6.2
GnuPGInterface==0.3.2
Jinja2==2.6
Mako==0.5.0
MarkupSafe==0.15
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
PyYAML==3.10
Pygments==1.4
Sphinx==1.1.3
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
adium-theme-ubuntu==0.3.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
beautifulsoup4==4.0.2
chardet==2.0.1
command-not-found==0.2.44
configglue==1.0
configobj==4.7.2
debtagshw==0.1
decorator==3.3.2
defer==1.0.2
dirspec==3.0.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
docutils==0.8.1
duplicity==0.6.18
elib.intl==0.0.3.dev
epydoc==3.0.1
gunicorn==18.0
httplib2==0.7.2
iotop==0.4.4
ipython==0.12.1
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
louis==2.3.0
lxml==2.3.2
mysql-connector-python==0.3.2-devel
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
paramiko==1.7.7.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psycopg2==2.5.2
pyOpenSSL==0.12
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pygeocoder==1.2.2
pyinotify==0.9.2
pyserial==2.5
pysmbc==1.0.13
pysqlite==2.6.3
pystache==0.5.3
python-apt==0.8.3ubuntu7.1
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-forecastio==0.1
python-virtkey==0.60.0
pytz==2011k
pyxdg==0.19
reportlab==2.5
requests==2.2.1
rhythmbox-ubuntuone==4.2.0
sessioninstaller==0.0.0
simplegeneric==0.7
simplejson==2.3.2
six==1.5.2
software-center-aptd-plugins==0.0.0
static==1.0.2
system-service==0.1.6
traits==4.0.0
twilio==3.6.5
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
wadllib==1.3.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==2.5.3
xkit==0.0.0
xlwt==0.7.2
zope.interface==3.6.1
